requirements.txt
pyswisseph==2.10.3.2
timezonefinder==6.2.0
pytz==2023.3
